+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Version 17
+
+* Removed `commit` from plusone::net::mmap_rx::frame
+    Now it called in `frame` destructor
+
 # Version 16
 
 * plusone::serialization refactoring